finish task 14

+++ src/main/java/ops/gateway/model/dto/MakerSuccessDataDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.criteria.CriteriaBuilder;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -11,5 +13,5 @@ public class MakerSuccessDataDTO {
     private String originalTransId;
     private Long objectId;
     private String desc;
-    private double amount;
+    private Integer amount;
 }
